make module structure
